{"schema":"decodex/commit/1","summary":"Split Decodex app panel usage modules","authority":"manual"}
