+++ apps/decodex-app/Sources/DecodexApp/AccountDisplayIdentity.swift
@@ -1,0 +1,58 @@
+import Foundation
+
+extension AccountDisplay {
+	static func compactEmail(_ email: String) -> String {
+		let text = email.trimmingCharacters(in: .whitespacesAndNewlines)
+		guard let atIndex = text.firstIndex(of: "@"), atIndex > text.startIndex else {
+			return compactIdentity(text)
+		}
+
+		let local = String(text[..<atIndex])
+		let domain = String(text[atIndex...])
+		if local.count <= 6 {
+			return "\(local)\(domain)"
+		}
+
+		return "\(local.prefix(3))...\(compactLocalSuffix(local))\(domain)"
+	}
+
+	static func compactIdentity(_ value: String) -> String {
+		let text = trimLeadingEllipsis(value)
+		if text.isEmpty || text == "unknown" {
+			return text
+		}
+
+		let edgeLength = max(3, min(6, text.count / 2))
+		return "\(text.prefix(edgeLength))...\(text.suffix(edgeLength))"
+	}
+
+	static func trimLeadingEllipsis(_ value: String) -> String {
+		let text = value.trimmingCharacters(in: .whitespacesAndNewlines)
+		if text.hasPrefix("..."), text.dropFirst(3).contains("...") == false {
+			return String(text.dropFirst(3))
+		}
+
+		return text
+	}
+
+	static func compactLocalSuffix(_ local: String) -> String {
+		if let separator = local.lastIndex(of: ".") {
+			let segment = String(local[local.index(after: separator)...])
+			if (2...4).contains(segment.count), segment.allSatisfy(\.isLetter) {
+				return segment
+			}
+		}
+
+		return String(local.suffix(3))
+	}
+
+	static func identityHash(_ value: String) -> UInt32 {
+		var hash: UInt32 = 2_166_136_261
+		for unit in value.utf16 {
+			hash ^= UInt32(unit)
+			hash = hash &* 16_777_619
+		}
+
+		return hash
+	}
+}
