+++ apps/decodex-app/Sources/DecodexApp/AccountUsageMeterView.swift
@@ -8,7 +8,7 @@ struct AccountUsageMeterView: View {
 	let dailyAveragePercent: Double?
 	let tone: AccountTone
 	let currentTime: Date
-	@Environment(\.colorScheme) private var colorScheme
+	@Environment(\.colorScheme) var colorScheme
 
 	var body: some View {
 		VStack(alignment: .leading, spacing: 3) {
@@ -91,100 +91,7 @@ struct AccountUsageMeterView: View {
 		}
 		.lineLimit(1)
 		.frame(height: 22)
-		.frame(maxWidth: .infinity, alignment: .leading)
-		.accessibilityLabel(accessibilityText)
-	}
-
-	private var remainingText: String {
-		guard let remainingPercent else {
-			return "-"
-		}
-
-		return "\(remainingPercent)% left"
-	}
-
-	private var dailyAverageText: String? {
-		guard let dailyAveragePercent else {
-			return nil
-		}
-		let formatted = formatDailyUsageRate(dailyAveragePercent)
-
-		return formatted == "-" ? nil : formatted
-	}
-
-	private var accessibilityText: String {
-		let average = dailyAverageText.map { ", daily average \($0)" } ?? ""
-		return "\(label) remaining \(remainingText)\(average), \(resetDisplay.accessibility)"
-	}
-
-	private var progress: CGFloat {
-		guard let remainingPercent else {
-			return 0
-		}
-
-		return CGFloat(max(0, min(100, remainingPercent))) / 100
-	}
-
-	private func fillWidth(in width: CGFloat) -> CGFloat {
-		guard remainingPercent != nil else {
-			return 0
-		}
-
-		return max(4, width * progress)
-	}
-
-	private var color: Color {
-		switch tone {
-		case .codexActive: return PanelPalette.codexAccent(colorScheme)
-		case .ready: return PanelPalette.capacityAccent(colorScheme)
-		case .selected: return PanelPalette.routeAccent(colorScheme)
-		case .warning: return PanelPalette.warning(colorScheme)
-		case .danger: return PanelPalette.destructive(colorScheme)
-		case .neutral: return PanelPalette.secondaryText(colorScheme)
-		}
-	}
-
-	private var valueColor: Color {
-		switch tone {
-		case .warning, .danger:
-			return color.opacity(colorScheme == .dark ? 0.95 : 0.78)
-		default:
-			return PanelPalette.primaryText(colorScheme).opacity(colorScheme == .dark ? 0.9 : 0.84)
-		}
-	}
-
-	private var resetDisplay: UsageResetDisplay {
-		UsageResetDisplay.make(resetAtUnixEpoch: resetAtUnixEpoch, now: currentTime)
-	}
-
-	private var trackColor: Color {
-		PanelPalette.progressTrack(colorScheme)
-	}
-
-	private var trackEdgeColor: Color {
-		PanelPalette.progressEdge(colorScheme)
-	}
-
-	private var fillStyle: LinearGradient {
-		LinearGradient(
-			colors: [
-				color.opacity(colorScheme == .dark ? 0.78 : 0.68),
-				color.opacity(colorScheme == .dark ? 0.62 : 0.52),
-			],
-			startPoint: .leading,
-			endPoint: .trailing
-		)
-	}
-
-	private var trackInsetStyle: LinearGradient {
-		LinearGradient(
-			colors: [
-				Color.white.opacity(colorScheme == .dark ? 0.022 : 0.05),
-				Color.white.opacity(0),
-				Color.black.opacity(colorScheme == .dark ? 0.035 : 0.018),
-			],
-			startPoint: .top,
-			endPoint: .bottom
-		)
+			.frame(maxWidth: .infinity, alignment: .leading)
+			.accessibilityLabel(accessibilityText)
 	}
 }
